fixed review comments
