+++ src/mirror.cpp
@@ -1,12 +1,40 @@
-#include <spdlog/fmt/fmt.h>
 
 #include <powerloader/mirror.hpp>
 #include <powerloader/mirrors/oci.hpp>
 #include "powerloader/mirrorid.hpp"
 #include "target.hpp"
 
+#include <fmt/core.h>
+
 namespace powerloader
 {
+    namespace
+    {
+        std::string strip_trailing_slash(const std::string& s)
+        {
+            if (s.size() > 0 && s.back() == '/' && s != "file://")
+            {
+                return s.substr(0, s.size() - 1);
+            }
+            return s;
+        }
+    }
+
+    MirrorID Mirror::id(const std::string& url)
+    {
+        return MirrorID(fmt::format("Mirror[{}]", url));
+    }
+
+    Mirror::Mirror(const MirrorID& id, const Context& ctx, const std::string& url)
+        : m_id(id)
+        , m_url(strip_trailing_slash(url))
+    {
+        if (ctx.max_downloads_per_mirror > 0)
+        {
+            m_stats.allowed_parallel_connections = ctx.max_downloads_per_mirror;
+        }
+    }
+
     Mirror::~Mirror() = default;
 
     void Mirror::change_max_ranges(int new_value)
@@ -183,6 +211,16 @@ namespace powerloader
         }
 
         return true;
+    }
+
+    HTTPMirror::HTTPMirror(const Context& ctx, const std::string& url)
+        : Mirror(HTTPMirror::id(url), ctx, url)
+    {
+    }
+
+    MirrorID HTTPMirror::id(const std::string& url)
+    {
+        return MirrorID{ fmt::format("HTTPMirror[{}]", url) };
     }
 
     bool HTTPMirror::authenticate(CURLHandle& handle, const std::string& path)
